Switched over to use actual files/fields
